requirements.txt
langchain 
wikipedia 
langchain-community 
python-dotenv
langchain-google-genai
pydantic
duckduckgo-search
langchain-openai
streamlit
tenacity